server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = "Tortuga app for digital health"
authors = ["Jean Nei <jeanbhs09@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
django = "^2.1"
djangorestframework = "^3.9"
django-cors-headers = "^2.5"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
pylint = "^2.3"
coreapi = "^2.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
